requirements.txt

python-binance